Implement Battles

+++ app/src/app/campaign/section/section.component.ts
@@ -1,49 +1,37 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from '@angular/core';
-import { Subject, take } from 'rxjs';
+import { BehaviorSubject, debounceTime, distinct, distinctUntilChanged, filter, Subject, switchMap, take, takeUntil, tap } from 'rxjs';
 import { CampaignSection } from 'src/app/interfaces/campaign-section';
+import { ConfirmDialogComponent, ConfirmDialogInterface } from 'src/app/misc/dialogs/confirm-dialog/confirm-dialog.component';
 import { ModalService } from 'src/app/services/modal.service';
 import { StoreService } from 'src/app/services/store.service';
 
 @Component({
   selector: 'campaign-section',
   template: `
-  <div class="flex flex-row" (click)="clicked()">
-    <input *ngIf="editing; else display" class="grow text-white p-2 m-2 rounded-md placeholder:text-slate-400 bg-dark-input-bg"
-      [(ngModel)]="section.name"
-      (input)="dirty = true"
-      placeholder="New Section"
-    >
-    <ng-template #display>
-      <div class="grow text-white p-2 m-2 cursor-pointer"
-        (dblclick)="editing = !editing">
-        {{section.name}}
-      </div>
-    </ng-template>
-    <button
-      *ngIf="dirty"
-      class="p-2 m-2 rounded-md text-white bg-dark-action hover:bg-dark-action-hover"
-      (click)="save()"
-      i18n i18n-title title="Save Section">
-      <img class="w-[28px] h-[28px]" src="assets/save-white.png" i18n-title title="Save" alt="Save"/>
-    </button>
-    <button
-      *ngIf="section.id"
-      class="p-2 m-2 rounded-md text-white bg-dark-action hover:bg-dark-accent-red"
-      (click)="onDeleteClicked()"
-      i18n i18n-title title="Delete Section">
-      <img class="w-[28px] h-[28px]" src="assets/delete-white.png" i18n-title title="Delete" alt="Delete"/>
-    </button>
+  <div class="flex flex-row w-full" (click)="clicked()">
+    <cInput
+      class="grow m-2"
+      [value]="section.name"
+      (valueChange)="title$.next($event)"
+      placeholder="Section Title"
+      title="Section Title"></cInput>
+    @if(section.id) {
+    <delete-button (click)="onDeleteClicked()" title="Delete Section"></delete-button>
+    }
   </div>
   `,
   styles: []
 })
-export class SectionComponent implements OnInit {
+export class SectionComponent implements OnInit, OnDestroy {
   @Input() section!: CampaignSection;
+  @Input() index!: number;
   @Output() sectionChange = new EventEmitter<CampaignSection>();
   @Output() selected = new EventEmitter<CampaignSection>();
+  @Output() deleteClicked = new EventEmitter<number>();
 
-  editing = false;
-  dirty = false;
+  title$ = new BehaviorSubject<string>('');
+
+  destroy$ = new Subject<void>();
 
   constructor(
     private store: StoreService,
@@ -51,15 +39,24 @@ export class SectionComponent implements OnInit {
   ) {}
   
   ngOnInit(): void {
-    if (this.section && this.section.name == "") {
-      this.editing = true;
+    if (this.section.id) {
+      this.title$.next(this.section.name);
     }
+    this.title$.pipe(
+      debounceTime(450),
+      distinctUntilChanged(),
+      filter(title => title != ""),
+      switchMap(title => {
+        this.section.name = title;
+        return this.store.saveSection(this.section);
+      }),
+      takeUntil(this.destroy$)
+    ).subscribe(s => this.section = s);
   }
 
-  save(): void {
-    this.store.saveSection(this.section).pipe(
-      take(1)
-    ).subscribe(s => {this.section = s; this.selected.emit(this.section);});
+  ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.complete();
   }
 
   clicked(): void {
@@ -68,13 +65,21 @@ export class SectionComponent implements OnInit {
 
   onDeleteClicked() {
     if (this.section.id) {
-      this.modal.open({
-        header: "Are you sure?",
+      let data: ConfirmDialogInterface = {
         message: "Are you sure you want to delete this section?\nThis will delete all entries as well.",
         confirm: true,
         yes: () => this.store.deleteSection(this.section).pipe(
+          tap(_ => this.deleteClicked.emit(this.index)),
           take(1)
-        ).subscribe()
+        ).subscribe(s => this.modal.close()),
+        no: () => this.modal.close(),
+        ok: () => this.modal.close(),
+        cancel: () => this.modal.close(),
+      };
+      this.modal.open({
+        header: "Are you sure?",
+        component: ConfirmDialogComponent,
+        data: data
       });
     }
   }
